backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
PyPDF2==3.0.1
python-multipart==0.0.6
gunicorn==21.2.0
